Implementation of the actuarial life table method.

Implementation of the actuarial life table method.

lifetable(data, breaks = NULL, time_var = "time", status_var = "status",
  right_closed = FALSE, max_end = FALSE)

Arguments

data
Data set from which time and status variables will be extracted.

breaks
The time points of the interval borders.

time_var
The variable storing event times.

status_var
The variable storing the event indicator.

right_closed
Logical. If TRUE (default), intervals are assumed right_closed closed and left open. If FALSE left closed and right_closed open.

max_end
logical. Should the last interval span until the last observed censoring or event time (if larger than the largest specified cut point).

Value

A tibble containing riskset information as well as hazard and survival function calculated by using the actuarial life table method.

See also

split_info

Examples

data("Melanoma", package="MASS")
Melanoma$status <- ifelse(Melanoma$status == 1, 1, 0)
lifetable(Melanoma, breaks = seq(0, 6000, 1500))

#> # A tibble: 4 x 9
#>   tstart  tend    interval     n events dropouts riskset     hazard  survival
#>    <dbl> <dbl>       <chr> <int>  <dbl>    <dbl>   <dbl>      <dbl>     <dbl>
#> 1      0  1500    [0,1500)   205     36       10     200 0.18000000 0.8200000
#> 2   1500  3000 [1500,3000)   159     19       86     116 0.16379310 0.6856897
#> 3   3000  4500 [3000,4500)    54      2       48      30 0.06666667 0.6399770
#> 4   4500  6000 [4500,6000)     4      0        4       2 0.00000000 0.6399770