Implementation of the actuarial life table method.

Implementation of the actuarial life table method.

lifetable(data, formula, cut = NULL)

Arguments

data
a data frame

formula
a model formula

cut
the vector of timepoints to cut at

Value

A tibble containing riskset information as well as hazard and survival function calculated by using the actuarial life table method.

See also

split_info

Examples

data("Melanoma", package="MASS")
Melanoma$status <- ifelse(Melanoma$status == 1, 1, 0)
lifetable(Melanoma, Surv(time,status)~1, cut = seq(0, 6000, 1500))

#> # A tibble: 4 x 9
#>   tstart  tend interval        n events dropouts riskset hazard survival
#>    <dbl> <dbl> <chr>       <int>  <dbl>    <dbl>   <dbl>  <dbl>    <dbl>
#> 1      0  1500 (0,1500]      205  36.0     10.0   200    0.180     0.820
#> 2   1500  3000 (1500,3000]   159  19.0     86.0   116    0.164     0.686
#> 3   3000  4500 (3000,4500]    54   2.00    48.0    30.0  0.0667    0.640
#> 4   4500  6000 (4500,6000]     4   0        4.00    2.00 0         0.640